페이스북 관련 죽는 현상 수정

+++ src/com/bulgogi/recipe/http/model/Like.java
@@ -12,7 +12,7 @@ public class Like {
 	public int postId;
 	
 	@JsonProperty("fb_id")
-	public int facebookId;
+	public long facebookId;
 		
 	@Override
 	public String toString() {
